# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Healthcare Platform API",
    "description": "Enterprise Healthcare Platform Foundation API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/encounters": {
      "post": {
        "tags": [
          "Encounters"
        ],
        "summary": "Create a clinical encounter",
        "description": "Create a new clinical encounter directly.",
        "operationId": "create_encounter_api_v1_encounters_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EncounterCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EncounterResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/encounters/{encounter_id}": {
      "get": {
        "tags": [
          "Encounters"
        ],
        "summary": "Get encounter details",
        "description": "Retrieve details of a clinical encounter by UUID.",
        "operationId": "get_encounter_api_v1_encounters__encounter_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "encounter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Encounter Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EncounterResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Encounters"
        ],
        "summary": "Update clinical notes or chief complaint for an in-progress encounter",
        "description": "Update active encounter notes. Completed encounters are locked against direct edits.",
        "operationId": "update_encounter_api_v1_encounters__encounter_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "encounter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Encounter Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EncounterUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EncounterResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/patients/{patient_id}/encounters": {
      "get": {
        "tags": [
          "Encounters"
        ],
        "summary": "Get longitudinal encounter history for a patient",
        "description": "Retrieve paginated historical clinical encounters for a patient.",
        "operationId": "get_patient_encounters_api_v1_patients__patient_id__encounters_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Patient Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EncounterListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/encounters/{encounter_id}/complete": {
      "post": {
        "tags": [
          "Encounters"
        ],
        "summary": "Complete a clinical encounter",
        "description": "Finalize clinical encounter, set completion timestamp, and lock the record.",
        "operationId": "complete_encounter_api_v1_encounters__encounter_id__complete_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "encounter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Encounter Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/EncounterCompleteRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EncounterResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/encounters/{encounter_id}/vitals": {
      "post": {
        "tags": [
          "Encounters"
        ],
        "summary": "Record observation vitals for an encounter",
        "description": "Record a vitals snapshot (BP, HR, SpO2, Temp, RR) for an active clinical encounter.",
        "operationId": "record_vitals_api_v1_encounters__encounter_id__vitals_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "encounter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Encounter Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VitalCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VitalResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Encounters"
        ],
        "summary": "Get all recorded vitals observations for an encounter",
        "description": "Retrieve time-series vitals observations recorded during an encounter.",
        "operationId": "get_encounter_vitals_api_v1_encounters__encounter_id__vitals_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "encounter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Encounter Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VitalListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EncounterCompleteRequest": {
        "properties": {
          "clinical_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clinical Notes"
          }
        },
        "type": "object",
        "title": "EncounterCompleteRequest"
      },
      "EncounterCreate": {
        "properties": {
          "patient_id": {
            "type": "string",
            "format": "uuid",
            "title": "Patient Id"
          },
          "provider_id": {
            "type": "integer",
            "title": "Provider Id"
          },
          "facility_id": {
            "type": "string",
            "format": "uuid",
            "title": "Facility Id"
          },
          "appointment_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Appointment Id"
          },
          "encounter_type": {
            "$ref": "#/components/schemas/EncounterTypeEnum",
            "default": "OUTPATIENT"
          },
          "chief_complaint": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Chief Complaint"
          },
          "clinical_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clinical Notes"
          }
        },
        "type": "object",
        "required": [
          "patient_id",
          "provider_id",
          "facility_id"
        ],
        "title": "EncounterCreate"
      },
      "EncounterListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EncounterResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "total_pages": {
            "type": "integer",
            "title": "Total Pages"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "total_pages"
        ],
        "title": "EncounterListResponse"
      },
      "EncounterResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "patient_id": {
            "type": "string",
            "format": "uuid",
            "title": "Patient Id"
          },
          "appointment_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Appointment Id"
          },
          "provider_id": {
            "type": "integer",
            "title": "Provider Id"
          },
          "facility_id": {
            "type": "string",
            "format": "uuid",
            "title": "Facility Id"
          },
          "encounter_type": {
            "$ref": "#/components/schemas/EncounterTypeEnum"
          },
          "status": {
            "$ref": "#/components/schemas/EncounterStatusEnum"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "chief_complaint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chief Complaint"
          },
          "clinical_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clinical Notes"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "updated_by": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated By"
          }
        },
        "type": "object",
        "required": [
          "id",
          "patient_id",
          "provider_id",
          "facility_id",
          "encounter_type",
          "status",
          "started_at",
          "created_at",
          "updated_at"
        ],
        "title": "EncounterResponse"
      },
      "EncounterStatusEnum": {
        "type": "string",
        "enum": [
          "SCHEDULED",
          "IN_PROGRESS",
          "COMPLETED",
          "CANCELLED"
        ],
        "title": "EncounterStatusEnum"
      },
      "EncounterTypeEnum": {
        "type": "string",
        "enum": [
          "OUTPATIENT",
          "EMERGENCY",
          "INPATIENT",
          "FOLLOW_UP",
          "HOME_VISIT"
        ],
        "title": "EncounterTypeEnum"
      },
      "EncounterUpdate": {
        "properties": {
          "chief_complaint": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Chief Complaint"
          },
          "clinical_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clinical Notes"
          }
        },
        "type": "object",
        "title": "EncounterUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VitalCreate": {
        "properties": {
          "temperature": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 45.0,
                "minimum": 30.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature",
            "description": "Temperature in Celsius (\u00b0C)"
          },
          "heart_rate": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 260.0,
                "minimum": 20.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Heart Rate",
            "description": "Heart rate in beats per minute (bpm)"
          },
          "respiratory_rate": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 80.0,
                "minimum": 5.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Respiratory Rate",
            "description": "Respiratory rate in breaths/min"
          },
          "systolic_bp": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 300.0,
                "minimum": 50.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Systolic Bp",
            "description": "Systolic blood pressure in mmHg"
          },
          "diastolic_bp": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 200.0,
                "minimum": 30.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Diastolic Bp",
            "description": "Diastolic blood pressure in mmHg"
          },
          "spo2": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 40.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Spo2",
            "description": "Blood oxygen saturation percentage (%)"
          },
          "weight": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 400.0,
                "minimum": 0.5
              },
              {
                "type": "null"
              }
            ],
            "title": "Weight",
            "description": "Weight in kilograms (kg)"
          },
          "height": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 280.0,
                "minimum": 20.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Height",
            "description": "Height in centimeters (cm)"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "title": "VitalCreate",
        "description": "Clinical observations and vitals recording payload.\n\nEnforces physiological sanity checks without replacing clinical diagnosis."
      },
      "VitalListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/VitalResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "encounter_id": {
            "type": "string",
            "format": "uuid",
            "title": "Encounter Id"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "encounter_id",
          "total"
        ],
        "title": "VitalListResponse"
      },
      "VitalResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "encounter_id": {
            "type": "string",
            "format": "uuid",
            "title": "Encounter Id"
          },
          "patient_id": {
            "type": "string",
            "format": "uuid",
            "title": "Patient Id"
          },
          "recorded_by": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recorded By"
          },
          "recorded_at": {
            "type": "string",
            "format": "date-time",
            "title": "Recorded At"
          },
          "temperature": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature"
          },
          "heart_rate": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Heart Rate"
          },
          "respiratory_rate": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Respiratory Rate"
          },
          "systolic_bp": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Systolic Bp"
          },
          "diastolic_bp": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Diastolic Bp"
          },
          "spo2": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spo2"
          },
          "weight": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weight"
          },
          "height": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "id",
          "encounter_id",
          "patient_id",
          "recorded_at"
        ],
        "title": "VitalResponse"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}